preliminary coding and styling of home.php
Having issues with the CSS of the home.php's header. Header does not
align properly as .header_buttons falls below the alignment of the
other header items. .line divs are repeated to properly style for their
inclusion in .line_container however soon this process will be taken
over with php.